+++ src/main/java/VoIP/Client/LoginGUI.java
@@ -12,86 +12,241 @@ public class LoginGUI extends JFrame {
         listeners();
     }
 
+    public LoginGUI getGUI() {
+        return this;
+    }
+
     private void listeners() {
         butLogin.addMouseListener(new MouseAdapter() {
             public void mouseClicked(MouseEvent evt) {
-                System.out.println("Hello World!");
+                // TODO: Check if credialtials are there.
+                try {
+                /* Applies the system look and feel */
+                UIManager.setLookAndFeel(
+                    UIManager.getSystemLookAndFeelClassName());
+                } catch (UnsupportedLookAndFeelException e) {
+                   System.out.println("System look and feel unsupported.\n" + e);
+                } catch (ClassNotFoundException e) {
+                   System.out.println("System look and feel class could not be found.\n" + e);
+                } catch (InstantiationException e) {
+                   System.out.println("System look and feel could not be initialized.\n" + e);
+                } catch (IllegalAccessException e) {
+                   System.out.println("Illegal Access Exception.\n" + e);
+                }
+
+                /* Create and display the form */
+                java.awt.EventQueue.invokeLater(new Runnable() {
+                    public void run() {
+                        new ChatGUI();
+                        getGUI().dispose();
+                    }
+                });
+            }
+        });
+
+        butRegister.addMouseListener(new MouseAdapter() {
+            public void mouseClicked(MouseEvent evt) {
+                // TODO: Check if info is filled in correctly
+                new ChatGUI();
+                getGUI().dispose();
             }
         });
     }
 
     private void initComponents() {
-        // Initialise Components.
-        panBlock = new JPanel();
-        txfUsername = new JTextField();
-        txfIPAddress = new JTextField();
-        labIPAddress = new JLabel("Server IP Address :");
-        labUsername = new JLabel("Username :");
-        butLogin = new JButton("Login");
-        labError = new JLabel();
-        labHeader = new JLabel("Welcome to VoIP!");
-        labBG = new JLabel();
+        // Initialise Main Components.
+        UIManager.put("TabbedPane.contentOpaque", false);
+        labBackground   = new JLabel();
+        tabMain         = new JTabbedPane();
 
-        // Settings of Components
-        labError.setText("<Error Message goes here>"); //TEMP
-        labError.setForeground(new Color(255, 0, 0));
-        labError.setHorizontalAlignment(SwingConstants.CENTER);
+        // Initialise Login Components
+        panLogin        = new JPanel();
+        txfLogUsername  = new JTextField();
+        txfLogPassword  = new JTextField();
+        labLogUsername  = new JLabel();
+        labLogPassword  = new JLabel();
+        butLogin        = new JButton();
+        labLogError     = new JLabel();
+        labLogHeader    = new JLabel();
+        butFrgtPassword = new JButton();
+
+        // Initialise Register Components
+        panRegister     = new JPanel();
+        txfRegUsername  = new JTextField();
+        txfRegPassword  = new JTextField();
+        labRegUsername  = new JLabel();
+        labRegPassword  = new JLabel();
+        butRegister     = new JButton();
+        labRegError     = new JLabel();
+        labRegHeader   = new JLabel();
+        labRegEmail     = new JLabel();
+        txfRegEmail     = new JTextField();
+
+        // Settings of Main Components
         try {
-            labBG.setIcon(new ImageIcon(getClass().getResource("/LoginBG.jpg")));
+            labBackground.setIcon(new ImageIcon(getClass().getResource("/LoginBG.jpg")));
         } catch (Exception e) {
-            labBG.setIcon(new ImageIcon(getClass().getResource("../Images/LoginBG.jpg")));
+            labBackground.setIcon(new ImageIcon(getClass().getResource("../Images/LoginBG.jpg")));
         }
+        labBackground.setBounds(0, 0, 800, 600);
 
-        labHeader.setHorizontalAlignment(SwingConstants.CENTER);
-        labHeader.setFont(new Font("Yrsa SemiBold", 1, 48));
+        tabMain.addTab("Login", panLogin);
+        tabMain.addTab("Register", panRegister);
+        tabMain.setBounds(190, 160, 410, 270);
+        tabMain.setPreferredSize(new Dimension(400, 300));
+        tabMain.setBackground(new Color(255, 255, 255, 200));
 
-        panBlock.setBounds(200, 150, 400, 300);
-        panBlock.setSize(new Dimension(400, 300));
-        panBlock.setBackground(new Color(255, 255, 255, 180));
+        // Settings of Login Components
+        panLogin.setPreferredSize(new Dimension(400, 300));
+        panLogin.setBackground(new Color(255, 255, 255, 200));
 
-        // Panel layout.
-        GroupLayout panLayout = new GroupLayout(panBlock);
-        panBlock.setLayout(panLayout);
-        panLayout.setHorizontalGroup(
-            panLayout.createSequentialGroup()
+        labLogHeader.setFont(new Font("Yrsa SemiBold", 1, 48));
+        labLogHeader.setHorizontalAlignment(SwingConstants.CENTER);
+        labLogHeader.setText("Welcome Back");
+        labLogUsername.setText("Username: ");
+        labLogPassword.setText("Password: ");
+        labLogError.setForeground(new Color(255, 0, 0)); // RED
+        labLogError.setText("<Error Message goes here>"); //TEMP: To be removed.
+        labLogError.setHorizontalAlignment(SwingConstants.CENTER);
+
+        butLogin.setText("Login");
+        butFrgtPassword.setText("Forgot Password?");
+
+        // Settings of Login Components
+        panRegister.setPreferredSize(new Dimension(400, 300));
+        panRegister.setBackground(new Color(255, 255, 255, 200));
+
+        labRegHeader.setFont(new Font("Yrsa SemiBold", 1, 48));
+        labRegHeader.setHorizontalAlignment(SwingConstants.CENTER);
+        labRegHeader.setText("Welcome to VoIP!");
+        labRegUsername.setText("Username: ");
+        labRegEmail.setText("Email: ");
+        labRegPassword.setText("Password: ");
+        labRegError.setForeground(new Color(255, 0, 0));
+        labRegError.setHorizontalAlignment(SwingConstants.CENTER);
+        labRegError.setText("<Error Message goes here>");
+
+        butRegister.setText("Register");
+
+        // Panel Layouts
+        GroupLayout layLogin = new GroupLayout(panLogin);
+        panLogin.setLayout(layLogin);
+
+        layLogin.setHorizontalGroup(
+            layLogin.createSequentialGroup()
                 .addGap(25, 25, 25)
-                .addGroup(panLayout.createParallelGroup()
-                    .addComponent(labHeader)
-                    .addGroup(panLayout.createSequentialGroup()
-                        .addComponent(labIPAddress, 140, 140, 140)
-                        .addComponent(txfIPAddress, 210, 210, 210))
-                    .addGroup(panLayout.createSequentialGroup()
-                        .addComponent(labUsername, 140, 140, 140)
-                        .addComponent(txfUsername, 210, 210, 210))
-                    .addGroup(panLayout.createSequentialGroup()
-                        .addComponent(labError, 265, 265, 265)
-                        .addGap(10,10,10)
-                        .addComponent(butLogin, 75, 75, 75)))
+                .addGroup(layLogin.createParallelGroup(GroupLayout.Alignment.TRAILING)
+                    .addComponent(labLogHeader, 350, 350, 350)
+                    .addGroup(layLogin.createSequentialGroup()
+                        .addComponent(labLogError, 250, 250, 250)
+                        .addComponent(butLogin, 100, 100, 100))
+                    .addGroup(layLogin.createSequentialGroup()
+                        .addComponent(labLogPassword, 85, 85, 85)
+                        .addComponent(txfLogPassword, 265, 265, 265))
+                    .addGroup(layLogin.createSequentialGroup()
+                        .addComponent(labLogUsername, 85, 85, 85)
+                        .addComponent(txfLogUsername, 265, 265, 265))
+                    .addComponent(butFrgtPassword))
                 .addGap(25, 25, 25)
+        );
 
+        layLogin.setVerticalGroup(
+            layLogin.createSequentialGroup()
+                .addGap(25, 25, 25)
+                .addComponent(labLogHeader)
+                .addGroup(layLogin.createParallelGroup()
+                    .addComponent(labLogUsername)
+                    .addComponent(txfLogUsername))
+                .addGap(12, 12, 12)
+                .addGroup(layLogin.createParallelGroup()
+                    .addComponent(labLogPassword)
+                    .addComponent(txfLogPassword))
+                .addGap(12, 12, 12)
+                .addComponent(butFrgtPassword)
+                .addGap(12, 12, 12)
+                .addGroup(layLogin.createParallelGroup()
+                    .addComponent(labLogError)
+                    .addComponent(butLogin))
+                .addGap(25, 25, 25)
+            );
+
+
+        GroupLayout layRegister = new GroupLayout(panRegister);
+        panRegister.setLayout(layRegister);
+        layRegister.setHorizontalGroup(
+            layRegister.createSequentialGroup()
+                .addGap(25, 25, 25)
+                .addGroup(layRegister.createParallelGroup(GroupLayout.Alignment.TRAILING)
+                    .addComponent(labRegHeader, 350, 350, 350)
+                    .addGroup(layRegister.createSequentialGroup()
+                        .addComponent(labRegError, 250, 250, 250)
+                        .addComponent(butRegister, 100, 100, 100))
+                    .addGroup(layRegister.createSequentialGroup()
+                        .addComponent(labRegPassword, 85, 85, 85)
+                        .addComponent(txfRegPassword, 265, 265, 265))
+                    .addGroup(layRegister.createSequentialGroup()
+                        .addComponent(labRegUsername, 85, 85, 85)
+                        .addComponent(txfRegUsername, 265, 265, 265))
+                    .addGroup(layRegister.createSequentialGroup()
+                        .addComponent(labRegEmail, 85, 85, 85)
+                        .addComponent(txfRegEmail, 265, 265, 265)))
+                .addGap(25, 25, 25)
         );
-        panLayout.setVerticalGroup(
-            panLayout.createSequentialGroup()
-            .addGap(25,25,25)
-            .addComponent(labHeader)
-            .addGap(35,35,35)
-            .addGroup(panLayout.createParallelGroup()
-                .addComponent(labIPAddress, 25, 25, 25)
-                .addComponent(txfIPAddress, 25, 25, 25))
-            .addGap(20,20,20)
-            .addGroup(panLayout.createParallelGroup()
-                .addComponent(labUsername, 25, 25, 25)
-                .addComponent(txfUsername, 25, 25, 25))
-            .addGap(20,20,20)
-            .addGroup(panLayout.createParallelGroup()
-                .addComponent(labError, 25, 25, 25)
-                .addComponent(butLogin, 25, 25, 25))
-            .addGap(40,40,40)
-        );
+        /*
+        layRegister.setHorizontalGroup(
+            layRegister.createParallelGroup(GroupLayout.Alignment.LEADING)
+            .addGroup(layRegister.createSequentialGroup()
+                .addContainerGap()
+                .addGroup(layRegister.createParallelGroup(GroupLayout.Alignment.LEADING)
+                    .addGroup(layRegister.createSequentialGroup()
+                        .addComponent(labRegHeader, GroupLayout.PREFERRED_SIZE, 352, GroupLayout.PREFERRED_SIZE)
+                        .addContainerGap(26, Short.MAX_VALUE))
+                    .addGroup(layRegister.createSequentialGroup()
+                        .addGroup(layRegister.createParallelGroup(GroupLayout.Alignment.LEADING)
+                            .addGroup(layRegister.createParallelGroup(GroupLayout.Alignment.TRAILING, false)
+                                .addGroup(layRegister.createSequentialGroup()
+                                    .addComponent(labRegError, GroupLayout.DEFAULT_SIZE, GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE)
+                                    .addPreferredGap(LayoutStyle.ComponentPlacement.RELATED)
+                                    .addComponent(butRegister, GroupLayout.PREFERRED_SIZE, 75, GroupLayout.PREFERRED_SIZE))
+                                .addGroup(layRegister.createSequentialGroup()
+                                    .addComponent(labRegUsername)
+                                    .addPreferredGap(LayoutStyle.ComponentPlacement.RELATED)
+                                    .addComponent(txfRegUsername, GroupLayout.PREFERRED_SIZE, 214, GroupLayout.PREFERRED_SIZE)))
+                            .addGroup(layRegister.createSequentialGroup()
+                                .addComponent(labRegEmail)
+                                .addPreferredGap(LayoutStyle.ComponentPlacement.RELATED)
+                                .addComponent(txfRegEmail, GroupLayout.PREFERRED_SIZE, 214, GroupLayout.PREFERRED_SIZE))
+                            .addGroup(layRegister.createSequentialGroup()
+                                .addComponent(labRegPassword)
+                                .addPreferredGap(LayoutStyle.ComponentPlacement.RELATED)
+                                .addComponent(txfRegPassword, GroupLayout.PREFERRED_SIZE, 214, GroupLayout.PREFERRED_SIZE)))
+                        .addGap(0, 0, Short.MAX_VALUE))))
+        );*/
+        layRegister.setVerticalGroup(
+            layRegister.createSequentialGroup()
+                .addGap(25, 25, 25)
+                .addComponent(labRegHeader)
+                .addGroup(layRegister.createParallelGroup()
+                    .addComponent(labRegUsername)
+                    .addComponent(txfRegUsername))
+                .addGap(12, 12, 12)
+                .addGroup(layRegister.createParallelGroup()
+                    .addComponent(labRegPassword)
+                    .addComponent(txfRegPassword))
+                .addGap(12, 12, 12)
+                .addGroup(layRegister.createParallelGroup()
+                    .addComponent(labRegEmail)
+                    .addComponent(txfRegEmail))
+                .addGap(12, 12, 12)
+                .addGroup(layRegister.createParallelGroup()
+                    .addComponent(labRegError)
+                    .addComponent(butRegister))
+                .addGap(25, 25, 25)
+            );
 
         // Frame Settings
-        getContentPane().add(panBlock);
-        getContentPane().add(labBG);
+        getContentPane().add(tabMain);
+        getContentPane().add(labBackground);
         pack();
         setVisible(true);
         setLocationRelativeTo(null);
@@ -101,21 +256,28 @@ public class LoginGUI extends JFrame {
     }
 
     public static void main(String args[]) {
-        try {
-            UIManager.setLookAndFeel(UIManager.getSystemLookAndFeelClassName());
-        } catch (Exception e) {
-            System.out.println("Something went wrong with look and feel.");
-        }
         new LoginGUI();
     }
 
     private JButton butLogin;
-    private JLabel labBG;
-    private JLabel labIPAddress;
-    private JLabel labUsername;
-    private JLabel labError;
-    private JLabel labHeader;
-    private JPanel panBlock;
-    private JTextField txfUsername;
-    private JTextField txfIPAddress;
+    private JButton butRegister;
+    private JButton butFrgtPassword;
+    private JLabel labLogPassword;
+    private JLabel labRegPassword;
+    private JLabel labLogUsername;
+    private JLabel labRegUsername;
+    private JLabel labBackground;
+    private JLabel labLogHeader;
+    private JLabel labRegHeader;
+    private JLabel labLogError;
+    private JLabel labRegError;
+    private JLabel labRegEmail;
+    private JPanel panRegister;
+    private JPanel panLogin;
+    private JTextField txfLogUsername;
+    private JTextField txfLogPassword;
+    private JTextField txfRegUsername;
+    private JTextField txfRegPassword;
+    private JTextField txfRegEmail;
+    private JTabbedPane tabMain;
 }
